FK-221 - Force second column to be treated as string in Excel
Order number looks like a numeric value, and Excel therefor loads it as such. There is not enough floating point precision in Excel to represent the full value.

This changes that column to be treated as string instead, since the order number isn't really a number that should be used in calculations.